+++ include/AbstractEqualityPolyhedra.hpp
@@ -4,7 +4,7 @@
 
 template <class P, typename T>
 struct AbstractEqualityPolyhedra : public AbstractPolyhedra<P, T> {
-    Matrix<intptr_t, 0, 0, 0> E;
+    Matrix<int64_t, 0, 0, 0> E;
     llvm::SmallVector<T, 8> q;
 
     using AbstractPolyhedra<P, T>::A;
@@ -12,9 +12,9 @@ struct AbstractEqualityPolyhedra : public AbstractPolyhedra<P, T> {
     using AbstractPolyhedra<P, T>::removeVariable;
     using AbstractPolyhedra<P, T>::getNumVar;
     using AbstractPolyhedra<P, T>::pruneBounds;
-    AbstractEqualityPolyhedra(Matrix<intptr_t, 0, 0, 0> A,
+    AbstractEqualityPolyhedra(Matrix<int64_t, 0, 0, 0> A,
                               llvm::SmallVector<T, 8> b,
-                              Matrix<intptr_t, 0, 0, 0> E,
+                              Matrix<int64_t, 0, 0, 0> E,
                               llvm::SmallVector<T, 8> q)
         : AbstractPolyhedra<P, T>(std::move(A), std::move(b)), E(std::move(E)),
           q(std::move(q)) {}
@@ -36,24 +36,24 @@ struct AbstractEqualityPolyhedra : public AbstractPolyhedra<P, T> {
 };
 
 struct IntegerEqPolyhedra
-    : public AbstractEqualityPolyhedra<IntegerEqPolyhedra, intptr_t> {
+    : public AbstractEqualityPolyhedra<IntegerEqPolyhedra, int64_t> {
 
-    IntegerEqPolyhedra(Matrix<intptr_t, 0, 0, 0> A,
-                       llvm::SmallVector<intptr_t, 8> b,
-                       Matrix<intptr_t, 0, 0, 0> E,
-                       llvm::SmallVector<intptr_t, 8> q)
+    IntegerEqPolyhedra(Matrix<int64_t, 0, 0, 0> A,
+                       llvm::SmallVector<int64_t, 8> b,
+                       Matrix<int64_t, 0, 0, 0> E,
+                       llvm::SmallVector<int64_t, 8> q)
         : AbstractEqualityPolyhedra(std::move(A), std::move(b), std::move(E),
                                     std::move(q)){};
-    bool knownLessEqualZeroImpl(intptr_t x) const { return x <= 0; }
-    bool knownGreaterEqualZeroImpl(intptr_t x) const { return x >= 0; }
+    bool knownLessEqualZeroImpl(int64_t x) const { return x <= 0; }
+    bool knownGreaterEqualZeroImpl(int64_t x) const { return x >= 0; }
 };
 struct SymbolicEqPolyhedra
     : public AbstractEqualityPolyhedra<SymbolicEqPolyhedra, MPoly> {
     PartiallyOrderedSet poset;
 
-    SymbolicEqPolyhedra(Matrix<intptr_t, 0, 0, 0> A,
+    SymbolicEqPolyhedra(Matrix<int64_t, 0, 0, 0> A,
                         llvm::SmallVector<MPoly, 8> b,
-                        Matrix<intptr_t, 0, 0, 0> E,
+                        Matrix<int64_t, 0, 0, 0> E,
                         llvm::SmallVector<MPoly, 8> q,
                         PartiallyOrderedSet poset)
         : AbstractEqualityPolyhedra(std::move(A), std::move(b), std::move(E),
