+++ include/LinearDiophantine.hpp
@@ -2,18 +2,18 @@
 #include "./Math.hpp"
 #include <tuple>
 
-llvm::Optional<std::tuple<intptr_t, intptr_t>>
-linearDiophantine(intptr_t c, intptr_t a, intptr_t b) {
+llvm::Optional<std::tuple<int64_t, int64_t>>
+linearDiophantine(int64_t c, int64_t a, int64_t b) {
     if (c == 0) {
-        return std::make_tuple(intptr_t(0), intptr_t(0));
+        return std::make_tuple(int64_t(0), int64_t(0));
     } else if ((a | b) == 0) {
         return {};
     }
     auto [g, x, y] = gcdx(a, b);
-    intptr_t cDivG = g == 1 ? c : c / g;
+    int64_t cDivG = g == 1 ? c : c / g;
     if (cDivG * g == c) {
         // g = a*x + b*y;
-        intptr_t cDivG = c / g;
+        int64_t cDivG = c / g;
         return std::make_tuple(x * cDivG, y * cDivG);
     } else {
         return {};
@@ -21,15 +21,15 @@ linearDiophantine(intptr_t c, intptr_t a, intptr_t b) {
 }
 
 // d = a*x; x = d/a
-llvm::Optional<std::tuple<intptr_t>> linearDiophantine(intptr_t d,
-                                                       std::tuple<intptr_t> a) {
-    intptr_t a0 = std::get<0>(a);
+llvm::Optional<std::tuple<int64_t>> linearDiophantine(int64_t d,
+                                                       std::tuple<int64_t> a) {
+    int64_t a0 = std::get<0>(a);
     if (d == 0) {
-        return std::make_tuple(intptr_t(0));
+        return std::make_tuple(int64_t(0));
     } else if (a0 == 0) {
         return {};
     }
-    intptr_t x = d / a0;
+    int64_t x = d / a0;
     if (a0 * x == d) {
         return std::make_tuple(x);
     } else {
@@ -37,8 +37,8 @@ llvm::Optional<std::tuple<intptr_t>> linearDiophantine(intptr_t d,
     }
 }
 // d = a[0]*x + a[1]*y;
-llvm::Optional<std::tuple<intptr_t, intptr_t>>
-linearDiophantine(intptr_t d, std::tuple<intptr_t, intptr_t> a) {
+llvm::Optional<std::tuple<int64_t, int64_t>>
+linearDiophantine(int64_t d, std::tuple<int64_t, int64_t> a) {
     return linearDiophantine(d, std::get<0>(a), std::get<1>(a));
 }
 
@@ -55,20 +55,20 @@ template <typename Tuple, size_t N> auto pop_front(const Tuple &tuple, Val<N>) {
 }
 
 template <typename Tuple>
-llvm::Optional<Tuple> linearDiophantine(intptr_t d, Tuple a) {
-    intptr_t a0 = std::get<0>(a);
-    intptr_t a1 = std::get<1>(a);
+llvm::Optional<Tuple> linearDiophantine(int64_t d, Tuple a) {
+    int64_t a0 = std::get<0>(a);
+    int64_t a1 = std::get<1>(a);
     auto aRem = pop_front(a, Val<2>());
     if ((a0 | a1) == 0) {
         auto opt = linearDiophantine(d, aRem);
         if (opt.hasValue()) {
-            return std::tuple_cat(std::make_tuple(intptr_t(0), intptr_t(0)),
+            return std::tuple_cat(std::make_tuple(int64_t(0), int64_t(0)),
                                   opt.getValue());
         } else {
             return {};
         }
     }
-    intptr_t q = std::gcd(a0, a1);
+    int64_t q = std::gcd(a0, a1);
     // d == q*((a/q)*x + (b/q)*y) + ... == q*w + ...
     // solve the rest
     auto dio_dqc =
@@ -77,7 +77,7 @@ llvm::Optional<Tuple> linearDiophantine(intptr_t d, Tuple a) {
         return {};
     }
     auto t = dio_dqc.getValue();
-    intptr_t w = std::get<0>(t);
+    int64_t w = std::get<0>(t);
     // w == ((a0/q)*x + (a1/q)*y)
     auto o = linearDiophantine(w, a0 / q, a1 / q);
     if (!o.hasValue()) {
