+++ include/Orthogonalize.hpp
@@ -8,12 +8,12 @@
 
 // `B` is a transposed mirror in reduced form
 // it is used to check whether a new row is linearly independent.
-static bool addIndRow(PtrMatrix<intptr_t> A, const Stride &axis, size_t j) {
-    // std::ranges::fill(A.getRow(j), intptr_t(0));
+static bool addIndRow(PtrMatrix<int64_t> A, const Stride &axis, size_t j) {
+    // std::ranges::fill(A.getRow(j), int64_t(0));
     for (size_t i = 0; i < axis.size(); ++i) {
         VarID v = axis[i].second;
         if (v.isLoopInductionVariable()) {
-            if (llvm::Optional<intptr_t> c =
+            if (llvm::Optional<int64_t> c =
                     axis[i].first.getCompileTimeConstant()) {
                 A(j, v.getID()) = c.getValue();
                 continue;
@@ -23,13 +23,13 @@ static bool addIndRow(PtrMatrix<intptr_t> A, const Stride &axis, size_t j) {
     }
     return false;
 }
-static bool addIndRow(PtrMatrix<intptr_t> A, const Stride &axis, size_t j,
+static bool addIndRow(PtrMatrix<int64_t> A, const Stride &axis, size_t j,
                       size_t k) {
-    // std::ranges::fill(A.getRow(j), intptr_t(0));
+    // std::ranges::fill(A.getRow(j), int64_t(0));
     for (size_t i = 0; i < axis.size(); ++i) {
         VarID v = axis[i].second;
         if (v.isLoopInductionVariable()) {
-            if (llvm::Optional<intptr_t> c =
+            if (llvm::Optional<int64_t> c =
                     axis[i].first.getCompileTimeConstant()) {
                 IDType id = v.getID();
                 if (id >= k)
@@ -62,8 +62,8 @@ orthogonalize(llvm::SmallVectorImpl<ArrayReference *> const &ai) {
     for (auto a : ai) {
         numRow += a->dim();
     }
-    Matrix<intptr_t, 0, 0> S(numRow, numLoops);
-    // std::ranges::fill(S, intptr_t(0));
+    Matrix<int64_t, 0, 0> S(numRow, numLoops);
+    // std::ranges::fill(S, int64_t(0));
     size_t row = 0;
     for (auto a : ai) {
         for (auto &axis : (*a)) {
@@ -81,7 +81,7 @@ orthogonalize(llvm::SmallVectorImpl<ArrayReference *> const &ai) {
         // A'*L <= b
         // now, we have (A = alnp.aln->A, r = alnp.aln->r)
         // (A'*K)*J <= r
-        Matrix<intptr_t, 0, 0, 0> A;
+        Matrix<int64_t, 0, 0, 0> A;
         matmultn(A, K, alnp.A);
         std::shared_ptr<AffineLoopNest> alnNew =
             std::make_shared<AffineLoopNest>(std::move(A), alnp.b, alnp.poset);
